pyproject.toml
[tool.poetry]
name = "winterthur-app"
version = "0.1.0"
description = "Web application for the Denig Manuscript project"
authors = ["Jason Heppler <jason.heppler@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "winterthur_app"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-tailwind = "^3.5.0"
django-htmx = "^1.14.0"
django-extensions = "^3.2.1"
django-environ = "^0.10.0"
humanfriendly = "^10.0"
django-taggit = "^3.1.0"
taggit-selectize = "^2.11.0"
django-multiselectfield = "^0.1.12"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
django-admin-interface = "^0.24.2"
python-dotenv = "^1.0.0"
django-import-export = "^3.2.0"
django-allauth = "^0.54.0"
daphne = "^4.0.0"
django-dbml = "^0.6.0"
django-prose = "^1.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
djhtml = "^3.0.6"
pre-commit = "^3.2.1"
django-debug-toolbar = "^4.0.0"
pytest-django = "^4.5.2"
django-extensions = "^3.2.1"
django-coverage-plugin = "^3.0.0"
pylint-django = "^2.5.3"
sphinx = "^7.0.0"
sphinx-autobuild = "^2021.3.14"
pytest-sugar = "^0.9.7"


[tool.poetry.group.base.dependencies]
python-slugify = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.djlint]
profile="django"